feat(storage): migrate game data storage to Room database (#10)
This commit refactors the entire data persistence layer, moving from a combination of SharedPreferences and direct JSON file parsing to a robust Room database implementation. This provides a scalable, centralized, and performant single source of truth for all game-related data.

Key changes:
- Introduced Room for persistent data storage, creating GameEntity and MemoryMatrixStateEntity to represent database tables.
- Implemented DAOs (GameDao, MemoryMatrixStateDao) for structured data access and an AppDatabase class to manage the database instance.
- Implemented a one-time database population from games.json on the first app launch using a RoomDatabase.Callback.
- Added a data synchronization mechanism that checks the app's versionCode on startup. If the app has been updated, it re-syncs the game data from games.json to ensure all users have the latest information.
