+++ app/src/main/README.md
@@ -1,0 +1,90 @@
+
+```
+main
+├─ AndroidManifest.xml
+├─ assets
+│  └─ games.json
+├─ java
+│  └─ com
+│     └─ negk
+│        └─ lerna
+│           ├─ data
+│           │  ├─ db
+│           │  │  ├─ AppDatabase.kt
+│           │  │  ├─ dao
+│           │  │  │  ├─ GameDao.kt
+│           │  │  │  └─ MemoryMatrixStateDao.kt
+│           │  │  └─ entity
+│           │  │     ├─ GameEntity.kt
+│           │  │     └─ MemoryMatrixStateEntity.kt
+│           │  ├─ GamePreviews.kt
+│           │  ├─ GameRegistry.kt
+│           │  ├─ Graph.kt
+│           │  └─ repository
+│           │     └─ GameRepository.kt
+│           ├─ GameActivity.kt
+│           ├─ MainActivity.kt
+│           ├─ navigation
+│           │  └─ screen.kt
+│           ├─ PreGameActivity.kt
+│           └─ ui
+│              ├─ components
+│              │  ├─ BaseCard.kt
+│              │  ├─ Header.kt
+│              │  └─ Shimmer.kt
+│              ├─ game
+│              │  ├─ GameHUD.kt
+│              │  ├─ memory
+│              │  │  ├─ MemoryMatrixGame.kt
+│              │  │  └─ MemoryMatrixViewModel.kt
+│              │  └─ speed
+│              │     ├─ SpeedMatchGame.kt
+│              │     └─ SpeedMatchViewModel.kt
+│              ├─ screens
+│              │  ├─ GamesScreen.kt
+│              │  ├─ home
+│              │  │  ├─ components
+│              │  │  │  ├─ DailyProgramCard.kt
+│              │  │  │  └─ RecommendedGameCard.kt
+│              │  │  └─ HomeScreen.kt
+│              │  ├─ pre_game
+│              │  │  └─ PreGameScreen.kt
+│              │  ├─ ProfileScreen.kt
+│              │  └─ TestsScreen.kt
+│              └─ theme
+│                 ├─ Color.kt
+│                 ├─ Theme.kt
+│                 └─ Type.kt
+└─ res
+   ├─ drawable
+   │  ├─ brain_svgrepo_com.xml
+   │  ├─ ic_launcher_background.xml
+   │  ├─ ic_launcher_foreground.xml
+   │  └─ memory_matrix.jpg
+   ├─ mipmap-anydpi
+   │  ├─ ic_launcher.xml
+   │  └─ ic_launcher_round.xml
+   ├─ mipmap-hdpi
+   │  ├─ ic_launcher.webp
+   │  └─ ic_launcher_round.webp
+   ├─ mipmap-mdpi
+   │  ├─ ic_launcher.webp
+   │  └─ ic_launcher_round.webp
+   ├─ mipmap-xhdpi
+   │  ├─ ic_launcher.webp
+   │  └─ ic_launcher_round.webp
+   ├─ mipmap-xxhdpi
+   │  ├─ ic_launcher.webp
+   │  └─ ic_launcher_round.webp
+   ├─ mipmap-xxxhdpi
+   │  ├─ ic_launcher.webp
+   │  └─ ic_launcher_round.webp
+   ├─ values
+   │  ├─ colors.xml
+   │  ├─ strings.xml
+   │  └─ themes.xml
+   └─ xml
+      ├─ backup_rules.xml
+      └─ data_extraction_rules.xml
+
+```
