+++ app/src/main/java/com/negk/lerna/GameActivity.kt
@@ -12,7 +12,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
 import com.negk.lerna.ui.theme.LernaTheme
-import com.negk.lerna.data.GameRegistry
+import com.negk.lerna.ui.game.utils.GameRegistry
 
 class GameActivity : ComponentActivity() {
