+++ app/src/main/java/com/negk/lerna/data/db/dao/MemoryMatrixStateDao.kt
@@ -1,15 +1,16 @@
-package com.negk.lerna.ui.game.memory
+package com.negk.lerna.data.db.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import com.negk.lerna.data.db.entity.MemoryMatrixStateEntity
 
 @Dao
 interface MemoryMatrixStateDao {
     @Query("SELECT * FROM memory_matrix_state WHERE id = :id")
     suspend fun getState(id: String = "default_state"): MemoryMatrixStateEntity?
 
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    @Insert(onConflict = OnConflictStrategy.Companion.REPLACE)
     suspend fun saveState(state: MemoryMatrixStateEntity)
 }
