+++ app/src/main/java/com/negk/lerna/data/db/entity/GameEntity.kt
@@ -1,4 +1,4 @@
-package com.negk.lerna.data
+package com.negk.lerna.data.db.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
