+++ app/src/main/java/com/negk/lerna/data/model/Game.kt
@@ -1,0 +1,10 @@
+package com.negk.lerna.data.model
+
+data class Game(
+    val id: String,
+    val title: String,
+    val description: String,
+    val hasLevels: Boolean,
+    val levels: List<String>,
+    val difficulty: String
+)
