+++ app/src/main/java/com/negk/lerna/data/model/GameJson.kt
@@ -1,0 +1,10 @@
+package com.negk.lerna.data.model
+
+data class GameJson(
+    val id: String,
+    val title: String,
+    val description: String,
+    val cognitiveArea: String,
+    val instructions: String,
+    val hasLevels: Boolean
+)
